Split server.py — extract machines route domain (#589)
Move the 4 machines handlers (api_machine_status, api_machines,
api_add_machine, api_remove_machine) plus their private helpers
(_check_machine_with_ip, _resolve_hostname, _check_machine_status) verbatim
into agentwire/routes/machines.py as MachinesRoutesMixin +
register_machines_routes, following the #584 seam.

_get_machines_data stays on the base server — it's consumed by the
dashboard WS, not these routes. Pure relocation: route set invariant
(150-route snapshot green).

Closes #589

Built by dotdev.dev
